+++ pgacollector/PgStatStatements.py
@@ -1,6 +1,3 @@
-# Copyright (c) 2014, pganalyze Team <team@pganalyze.com>
-#  All rights reserved.
-
 import logging
 import re
 import sys
@@ -12,36 +9,39 @@ class PgStatStatements():
 
     def __init__(self, db):
         self.db = db
-        if db.version_numeric < 90200:
-            logger.error("To use pg_stat_statements you must have at least Postgres 9.2 or newer")
-            sys.exit(1)
+
+    def have_stats_helper(self):
+        query = """
+        SELECT 1 AS enabled
+          FROM pg_proc
+          JOIN pg_namespace ON (pronamespace = pg_namespace.oid)
+         WHERE nspname = 'pganalyze' AND proname = 'get_stat_statements'
+        """
+        return self.db.run_query(query) == [{"enabled": 1}]
 
     def fetch_queries(self):
-        columns = ["userid", "dbid",
-                   "calls", "rows", "total_time",
-                   "shared_blks_hit", "shared_blks_read", "shared_blks_dirtied", "shared_blks_written",
-                   "local_blks_hit", "local_blks_read", "local_blks_dirtied", "local_blks_written",
-                   "temp_blks_read", "temp_blks_written",
-                   "blk_read_time", "blk_write_time"]
+        query = "SELECT * FROM "
 
-        query = "SELECT query AS normalized_query"
-
-        # Generate list of fields we'e interested in
-        query += ", " + ", ".join(columns)
-
-        query += " FROM pg_stat_statements"
+        if self.have_stats_helper():
+            query += "pganalyze.get_stat_statements()"
+        else:
+            query += "pg_stat_statements"
 
         # We don't want our stuff in the statistics
         query += " WHERE query !~* '^%s'" % re.sub(r'([*/])', r'\\\1', self.db.querymarker)
         # Filter out queries we shouldn't see in the first place
         query += " AND query <> '<insufficient privilege>'"
+        # Filter out DEALLOCATE statements - they are not useful and consume space
+        query += " AND query NOT LIKE 'DEALLOCATE %'"
         # Only get queries from current database
         query += " AND dbid IN (SELECT oid FROM pg_database WHERE datname = current_database())"
 
         queries = []
 
         for row in self.db.run_query(query, False):
-            row['plans'] = []
+            del row['dbid']
+            del row['userid']
+
             queries.append(row)
 
         return queries